python-server/requirements.txt
grpcio
grpcio-tools
torch
transformers